requirements.txt
flask==3.0.0
pdfminer.six==20221105
python-docx==1.1.0
spacy==3.7.2
sentence-transformers==2.2.2
torch==2.1.0
werkzeug==3.0.1